refactor: improvements to migrations
