+++ src/Lib/StatCan.OrchardCore.Extensions/ContentBuilderExtensions.cs
@@ -9,6 +9,8 @@ using OrchardCore.Html.Models;
 using OrchardCore.Markdown.Fields;
 using OrchardCore.Markdown.Models;
 using OrchardCore.Markdown.Settings;
+using OrchardCore.Title.Models;
+using StatCan.OrchardCore.ContentFields.MultiValueTextField.Settings;
 
 namespace StatCan.OrchardCore.Extensions
 {
@@ -132,6 +134,20 @@ namespace StatCan.OrchardCore.Extensions
             );
         }
 
+        public static ContentPartDefinitionBuilder WithMultiValueTextField(this ContentPartDefinitionBuilder p, string name, string displayName, string position, ListValueOption[] options)
+        {
+            return p.WithField(name, field => field
+               .OfType("MultiValueTextField")
+               .WithDisplayName(displayName)
+               .WithEditor("PredefinedList")
+               .WithPosition(position)
+               .WithSettings(new MultiValueTextFieldEditorSettings
+               {
+                   Options = options
+               })
+            );
+        }
+
         public static ContentTypeDefinitionBuilder WithHtmlBody(this ContentTypeDefinitionBuilder t, string position)
         {
             return t.WithPart(nameof(HtmlBodyPart), p => {
@@ -177,6 +193,21 @@ namespace StatCan.OrchardCore.Extensions
             manager.AlterPartDefinition(name, p => p.WithDisplayName(name));
             manager.AlterTypeDefinition(name, t => t.Stereotype("Widget")
                .WithPart(name, p => p.WithPosition("0"))
+            );
+        }
+
+        public static ContentTypeDefinitionBuilder WithTitlePart(this ContentTypeDefinitionBuilder t, string position)
+        {
+            return t.WithPart(nameof(TitlePart), p => p
+                .WithPosition(position)
+            );
+        }
+
+        public static ContentTypeDefinitionBuilder WithTitlePart(this ContentTypeDefinitionBuilder t, string position, TitlePartOptions options = TitlePartOptions.Editable, string pattern = "")
+        {
+            return t.WithPart(nameof(TitlePart), p => p
+                .WithPosition(position)
+                .WithSettings(new TitlePartSettings() { Options = options, Pattern = pattern })
             );
         }
     }
