+++ src/Modules/StatCan.OrchardCore.ContentFields/MultiValueTextField/Settings/MultiValueTextFieldEditorSettings.cs
@@ -1,4 +1,5 @@
 using Newtonsoft.Json;
+using OrchardCore.ContentFields.Settings;
 
 namespace StatCan.OrchardCore.ContentFields.MultiValueTextField.Settings
 {
@@ -14,14 +15,5 @@ namespace StatCan.OrchardCore.ContentFields.MultiValueTextField.Settings
         Checkbox,
         Dropdown,
         Picker
-    }
-
-    public class ListValueOption
-    {
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("value")]
-        public string Value { get; set; }
     }
 }
