+++ src/Modules/StatCan.OrchardCore.ContentFields/MultiValueTextField/Settings/MultiValueTextFieldEditorSettingsDriver.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
 using Microsoft.Extensions.Localization;
 using Newtonsoft.Json;
+using OrchardCore.ContentFields.Settings;
 using OrchardCore.ContentManagement.Metadata.Models;
 using OrchardCore.ContentTypes.Editors;
 using OrchardCore.DisplayManagement.Views;
