+++ src/Modules/StatCan.OrchardCore.Hackathon/StatCan.OrchardCore.Hackathon.csproj
@@ -21,7 +21,6 @@
 
   <ItemGroup>
     <PackageReference Include="HtmlSanitizer" />
-    <PackageReference Include="OrchardCore.Title" Version="$(OrchardCoreVersion)" />
   </ItemGroup>
 
   <ItemGroup>
