+++ src/Themes/HackathonTheme/Migrations.cs
@@ -1,13 +1,10 @@
-using OrchardCore.ContentFields.Settings;
 using OrchardCore.ContentManagement.Metadata;
 using OrchardCore.ContentManagement.Metadata.Settings;
 using OrchardCore.Data.Migration;
 using OrchardCore.Title.Models;
-using OrchardCore.Flows.Models;
-using ListValueOption = StatCan.OrchardCore.ContentFields.MultiValueTextField.Settings.ListValueOption;
 using StatCan.OrchardCore.Extensions;
 using StatCan.OrchardCore.ContentFields.MultiValueTextField.Settings;
-
+using OrchardCore.ContentFields.Settings;
 
 namespace StatCan.Themes.HackathonTheme
 {
@@ -25,6 +22,9 @@ namespace StatCan.Themes.HackathonTheme
             Tabs();
             ScheduleEvent();
             MenuItems();
+            VRow();
+            VContainer();
+            VCol();
             return 1;
         }
 
@@ -39,10 +39,7 @@ namespace StatCan.Themes.HackathonTheme
             );
 
             _contentDefinitionManager.AlterPartDefinition("HackathonThemeSettings", part => part
-                .WithField("DisplayName", field => field
-                    .OfType("TextField")
-                    .WithDisplayName("DisplayName")
-                    .WithPosition("0"))
+                .WithTextField("DisplayName", "0")
                 .WithField("Logo", field => field
                     .OfType("MediaField")
                     .WithDisplayName("Logo")
@@ -55,17 +52,12 @@ namespace StatCan.Themes.HackathonTheme
         {
             _contentDefinitionManager.CreateBasicWidget("Tabs");
             _contentDefinitionManager.AlterTypeDefinition("Tabs", t => t
-                .WithPart(nameof(TitlePart), p => p
-                    .WithPosition("0")
-                )
+                .WithTitlePart("0")
                 .WithFlow("1", new string[] { "Tab" })
             );
             _contentDefinitionManager.CreateBasicWidget("Tab");
             _contentDefinitionManager.AlterTypeDefinition("Tab", t => t
-                .WithPart(nameof(TitlePart), p => p
-                    .WithPosition("0")
-                    .WithSettings(new TitlePartSettings() { Options = TitlePartOptions.EditableRequired })
-                )
+                .WithTitlePart("0", TitlePartOptions.EditableRequired)
                 .WithFlow("2")
             );
         }
@@ -74,9 +66,7 @@ namespace StatCan.Themes.HackathonTheme
         {
             _contentDefinitionManager.CreateBasicWidget("ScheduleList");
             _contentDefinitionManager.AlterTypeDefinition("ScheduleList", t => t
-                .WithPart(nameof(TitlePart), p => p
-                    .WithPosition("0")
-                )
+                .WithTitlePart("0")
                 .WithFlow("1", new string[] { "ScheduleEvent" })
             );
             _contentDefinitionManager.AlterPartDefinition("ScheduleEvent", p => p
@@ -85,10 +75,7 @@ namespace StatCan.Themes.HackathonTheme
             );
             _contentDefinitionManager.AlterTypeDefinition("ScheduleEvent", t => t
                 .Stereotype("Widget")
-                .WithPart(nameof(TitlePart), p => p
-                    .WithPosition("0")
-                    .WithSettings(new TitlePartSettings() { Options = TitlePartOptions.EditableRequired })
-                )
+                .WithTitlePart("0", TitlePartOptions.EditableRequired)
                 .WithPart("ScheduleEvent", p => p.WithPosition("1"))
                 .WithMarkdownBody("2")
             );
@@ -97,18 +84,10 @@ namespace StatCan.Themes.HackathonTheme
         private void MenuItems()
         {
             _contentDefinitionManager.AlterPartDefinition("ContentMenuItem", part => part
-                .WithField("IconName", field => field
-                    .OfType("TextField")
-                    .WithDisplayName("Icon Name")
-                    .WithPosition("0")
-                )
+                .WithTextField("IconName","Icon Name" , "0")
             );
             _contentDefinitionManager.AlterPartDefinition("LinkMenuItem", part => part
-                .WithField("IconName", field => field
-                    .OfType("TextField")
-                    .WithDisplayName("Icon Name")
-                    .WithPosition("0")
-                )
+                .WithTextField("IconName", "Icon Name", "0")
             );
         }
 
@@ -120,56 +99,45 @@ namespace StatCan.Themes.HackathonTheme
                 .WithPart("VCol", part => part
                     .WithPosition("0")
                 )
-                .WithPart("FlowPart", part => part
-                    .WithPosition("1")
-                )
+                .WithFlow("1")
             );
 
             _contentDefinitionManager.AlterPartDefinition("VCol", part => part
-                .WithField("Props", field => field
-                    .OfType("MultiValueTextField")
-                    .WithDisplayName("Props")
-                    .WithEditor("PredefinedList")
-                    .WithPosition("0")
-                    .WithSettings(new MultiValueTextFieldEditorSettings
-                    {
-                        Options = new ListValueOption[] { new ListValueOption() {
-                    Name="Align Self Start",
-                    Value="align-self=\"start\""
+                .WithMultiValueTextField("Props","Props","0", new ListValueOption[] { new ListValueOption() {
+                        Name="Align Self Start",
+                        Value="align-self=\"start\""
                     }, new ListValueOption() {
-                    Name="Align Self Center",
-                    Value="align-self=\"center\""
+                        Name="Align Self Center",
+                        Value="align-self=\"center\""
                     }, new ListValueOption() {
-                    Name="Align Self End",
-                    Value="align-self=\"end\""
+                        Name="Align Self End",
+                        Value="align-self=\"end\""
                     }, new ListValueOption() {
-                    Name="Align Self Auto",
-                    Value="align-self=\"auto\""
+                        Name="Align Self Auto",
+                        Value="align-self=\"auto\""
                     }, new ListValueOption() {
-                    Name="Align Self Baseline",
-                    Value="align-self=\"baseline\""
+                        Name="Align Self Baseline",
+                        Value="align-self=\"baseline\""
                     }, new ListValueOption() {
-                    Name="Align Self Stretch",
-                    Value="align-self=\"stretch\""
+                        Name="Align Self Stretch",
+                        Value="align-self=\"stretch\""
                     }, new ListValueOption() {
-                    Name="Col 12 (Full)",
-                    Value="md=\"12\""
+                        Name="Col 12 (Full)",
+                        Value="md=\"12\""
                     }, new ListValueOption() {
-                    Name="Col 8 (3/4)",
-                    Value="md=\"8\""
+                        Name="Col 8 (3/4)",
+                        Value="md=\"8\""
                     }, new ListValueOption() {
-                    Name="Col 6 (1/2)",
-                    Value="md=\"6\""
+                        Name="Col 6 (1/2)",
+                        Value="md=\"6\""
                     }, new ListValueOption() {
-                    Name="Col 4 (1/4)",
-                    Value="md=\"4\""
+                        Name="Col 4 (1/4)",
+                        Value="md=\"4\""
                     } }
-                    })
                 )
             );
-
-
         }
+
         private void VRow()
         {
             _contentDefinitionManager.AlterTypeDefinition("VRow", type => type
@@ -178,49 +146,34 @@ namespace StatCan.Themes.HackathonTheme
                 .WithPart("VRow", part => part
                     .WithPosition("0")
                 )
-                .WithPart("FlowPart", part => part
-                    .WithPosition("1")
-                    .WithSettings(new FlowPartSettings
-                    {
-                        ContainedContentTypes = new[] { "VCol" },
-                    })
-                )
+                .WithFlow("1", new[] { "VCol" })
             );
 
             _contentDefinitionManager.AlterPartDefinition("VRow", part => part
-                .WithField("Props", field => field
-                    .OfType("MultiValueTextField")
-                    .WithDisplayName("Props")
-                    .WithEditor("PredefinedList")
-                    .WithPosition("0")
-                    .WithSettings(new MultiValueTextFieldEditorSettings
-                    {
-                        Options = new ListValueOption[] {
-                            new ListValueOption() {
+            .WithMultiValueTextField("Props", "Props", "0", new ListValueOption[] {
+                    new ListValueOption() {
                         Name="No Gutters",
                         Value="no-gutters"
-                        }, new ListValueOption() {
+                    }, new ListValueOption() {
                         Name="Justify Start",
                         Value="justify=\"start\""
-                        }, new ListValueOption() {
+                    }, new ListValueOption() {
                         Name="Justify Center",
                         Value="justify=\"center\""
-                        }, new ListValueOption() {
+                    }, new ListValueOption() {
                         Name="Justify End",
                         Value="justify=\"end\""
-                        }, new ListValueOption() {
+                    }, new ListValueOption() {
                         Name="Justify Between",
                         Value="justify=\"space-between\""
-                        }, new ListValueOption() {
+                    }, new ListValueOption() {
                         Name="Justify Around",
                         Value="justify=\"space-around\""
-                        } }
-                    })
+                    }}
                 )
             );
-
-
         }
+
         private void VContainer()
         {
             _contentDefinitionManager.AlterTypeDefinition("VContainer", type => type
@@ -229,29 +182,16 @@ namespace StatCan.Themes.HackathonTheme
                 .WithPart("VContainer", part => part
                     .WithPosition("0")
                 )
-                .WithPart("FlowPart", part => part
-                    .WithPosition("1")
-                    .WithSettings(new FlowPartSettings
-                    {
-                        ContainedContentTypes = new[] { "VRow" },
-                    })
-                )
+                .WithFlow("1", new[] { "VRow" })
             );
 
             _contentDefinitionManager.AlterPartDefinition("VContainer", part => part
-                .WithField("Props", field => field
-                    .OfType("MultiValueTextField")
-                    .WithDisplayName("Props")
-                    .WithEditor("PredefinedList")
-                    .WithSettings(new MultiValueTextFieldEditorSettings
-                    {
-                        Options = new ListValueOption[] {
-                            new ListValueOption() {
-                            Name="Fluid",
-                            Value="fluid"
+                .WithMultiValueTextField("Props", "Props", "0", new ListValueOption[] {
+                                new ListValueOption() {
+                                Name="Fluid",
+                                Value="fluid"
+                            }
                         }
-                    }
-                    })
                 )
             );
         }
