+++ src/Themes/HackathonTheme/Views/FlowPart.liquid
@@ -1,0 +1,3 @@
+{% for widget in Model.FlowPart.Widgets %}
+  {{ widget | shape_build_display | shape_render }}
+{% endfor %}
